requirements.txt
appdirs==1.4.3
backcall==0.1.0
beautifulsoup4==4.6.3
bleach==2.1.3
btrfsutil==1.0.0
certifi==2018.8.24
chardet==3.0.4
chrome-gnome-shell==0.0.0
cupshelpers==1.0
decorator==4.3.0
entrypoints==0.2.3
evdev==1.0.0
flake8==3.5.0
gufw==18.4.0
html5lib==1.0.1
idna==2.7
ipykernel==4.8.2
ipython==6.5.0
ipython-genutils==0.1.0
ipywidgets==7.2.1
Jade-Application-Kit==1.3.5
jedi==0.12.1
Jinja2==2.10
jsonschema==2.6.0
jupyter-client==5.2.3
jupyter-console==5.2.0
jupyter-core==4.4.0
lensfun==0.3.2
lit==0.6.0.dev0
louis==3.6.0
lutris==0.4.18.1
lxml==4.2.4
MarkupSafe==1.0
mccabe==0.6.1
mistune==0.8.3
nbconvert==5.3.1
nbformat==4.4.0
notebook==5.5.0
npyscreen==4.10.5
numpy==1.15.0
oauthlib==2.1.0
packaging==17.1
pacman-mirrors==4.12.5
pandocfilters==1.4.2
parso==0.3.1
pexpect==4.6.0
pickleshare==0.7.4
Pillow==5.2.0
prompt-toolkit==1.0.15
ptyprocess==0.6.0
pycairo==1.17.0
pycodestyle==2.4.0
pycups==1.9.73
pycurl==7.43.0.2
pyflakes==2.0.0
Pygments==2.2.0
pygobject==3.28.3
pyparsing==2.2.0
PyQt5-sip==4.19.12
pyserial==3.4
pyside==1.2.4
PySocks==1.6.8
python-dateutil==2.7.3
pyxdg==0.26
PyYAML==3.13
pyzmq==17.1.0
requests==2.19.1
requests-oauthlib==1.0.0
rpm==4.14.2
scipy==1.1.0
selenium==3.14.0
Send2Trash==1.5.0
simplegeneric==0.8.1
sip==4.19.12
six==1.11.0
team==1.0
terminado==0.8.1
testpath==0.3.1
tornado==5.1
traitlets==4.3.2
tweepy==3.6.0
ufw==0.35
urllib3==1.23
virtualenv==16.0.0
wcwidth==0.1.7
webencodings==0.5.1
widgetsnbextension==3.4.0
youtube-dl==2018.8.4
